requirements.txt
amqp==5.3.1
arrow==1.4.0
asgiref==3.11.0
billiard==4.2.4
binaryornot==0.4.4
brotli==1.2.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron_descriptor==2.0.6
cssselect2==0.8.0
Django==6.0.1
django-browser-reload==1.21.0
django-celery-results==2.5.1
django-environ==0.12.0
django-extensions==4.1
django-tailwind==4.4.2
django-timezone-field==7.2.1
django-widget-tweaks==1.5.1
dnspython==2.8.0
docx2pdf==0.1.8
et_xmlfile==2.0.0
eventlet==0.40.4
fonttools==4.61.1
greenlet==3.3.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
prompt_toolkit==3.0.52
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
pypdf==6.6.2
pyphen==0.17.2
pytailwindcss==0.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-magic==0.4.27
python-slugify==8.0.4
pywin32==311
PyYAML==6.0.3
qrcode==8.2
redis==5.0.1
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.4.0
webencodings==0.5.1
zopfli==0.4.0
